Add files via upload
This is the very first version of the bot. At this stage, when sending the /start command, the bot automatically saves the user ID in the Telegram, his first and last name, which are specified in the profile and username, to a pre-created database (if there is no database, the bot will create it automatically). The user can update information about himself using the /register command. When entering this command, the bot prompts the user to enter a new first and last name. The user can also find out how it is recorded in the database using the /check command. When you enter this command, the bot will find the user in the database by his id in the Telegram and send the requested information.